requirements.txt
fastapi
pydantic
bcrypt
python-jose[cryptography]
python-dotenv
psycopg2-binary
uvicorn[standard]
sqlalchemy